Deal with some type and value issues

+++ get_netcdf.py
@@ -89,7 +89,7 @@ class Pixel(object):
     def GetDead(self, fpeaks, speaks):
         # get detector dead time parameter "tau"
         try:
-            self.tau = fpeaks / speaks
+            self.tau = fpeaks.astype(float) / speaks
         except ZeroDivisionError:
             self.tau = fpeaks
